Reverted to edit a60b0fbf-9607-4e44-929a-463d6a1bf8ff: "Update benefit element width
Increased the width of the benefit elements in the Why Choose section to enhance visibility and layout.
[skip gpt_engineer]"
