--- src/components/marketing/landing/FDASeal.tsx
@@ -1,11 +1,0 @@
-export const FDASeal = () => {
-  return (
-    <div className="flex justify-center mt-4">
-      <img
-        src="/lovable-uploads/21d49d91-1f9d-4587-83fd-c4e63b5a8a9d.png"
-        alt="FDA Approved Seal"
-        className="h-16 w-16"
-      />
-    </div>
-  );
-};
